+++ ArgumentResolver/ApiCollectionEntryDtoArgumentResolver.php
@@ -13,21 +13,29 @@ declare(strict_types=1);
 
 namespace MarfaTech\Bundle\ApiPlatformBundle\ArgumentResolver;
 
-use Exception;
 use Generator;
 use MarfaTech\Bundle\ApiPlatformBundle\Exception\ApiException;
 use MarfaTech\Bundle\ApiPlatformBundle\Factory\ApiDtoFactory;
 use MarfaTech\Bundle\ApiPlatformBundle\HttpFoundation\ApiRequest;
 use MarfaTech\Component\DtoResolver\Dto\CollectionDtoResolverInterface;
+use Psr\Log\LoggerAwareInterface;
+use Psr\Log\LoggerAwareTrait;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Controller\ArgumentValueResolverInterface;
 use Symfony\Component\HttpKernel\ControllerMetadata\ArgumentMetadata;
 use Symfony\Component\OptionsResolver\Exception\InvalidOptionsException;
+use Symfony\Component\Validator\Exception\ValidationFailedException;
+use Throwable;
 
+use function array_key_first;
 use function is_subclass_of;
+use function iterator_to_array;
+use function sprintf;
 
-class ApiCollectionEntryDtoArgumentResolver implements ArgumentValueResolverInterface
+class ApiCollectionEntryDtoArgumentResolver implements ArgumentValueResolverInterface, LoggerAwareInterface
 {
+    use LoggerAwareTrait;
+
     /**
      * @var ApiDtoFactory
      */
@@ -59,14 +67,38 @@ class ApiCollectionEntryDtoArgumentResolver implements ArgumentValueResolverInte
      */
     public function resolve(Request $request, ArgumentMetadata $argument): Generator
     {
-        /** @var CollectionDtoResolverInterface $resolvedArgument */
         $resolvedArgument = $this->factory->createApiCollectionDto($argument->getType());
 
         foreach ($request->body->all() as $item) {
             try {
                 $resolvedArgument->add($item);
-            } catch (Exception|InvalidOptionsException $e) {
+            } catch (ValidationFailedException $e) {
+                $violationList = iterator_to_array($e->getViolations());
+                $violation = $violationList[array_key_first($violationList)];
+
+                $propertyPath = $violation->getPropertyPath();
+                $invalidValue = $violation->getInvalidValue();
+                $root = $violation->getRoot();
+
+                if ($propertyPath || $root || $invalidValue) {
+                    $message = sprintf(
+                        '%s: %s',
+                        $propertyPath ?: $root ?: $invalidValue,
+                        $violation->getMessage()
+                    );
+                } else {
+                    $message = $violation->getMessage();
+                }
+
+                throw new ApiException(ApiException::HTTP_BAD_REQUEST_DATA, $message);
+            } catch (InvalidOptionsException $e) {
                 throw new ApiException(ApiException::HTTP_BAD_REQUEST_DATA, $e->getMessage());
+            } catch (Throwable $e) {
+                if ($this->logger) {
+                    $this->logger->notice('Unexpected error while argument resolving', [$e]);
+                }
+
+                throw new ApiException(ApiException::HTTP_BAD_REQUEST_DATA);
             }
         }